+++ src/span.c
@@ -14,7 +14,7 @@ SPN_Span SPN_from_cstr(const char * cstr) {
   return (SPN_Span){.begin = (const void *)cstr, .len = strlen(cstr), .element_size = 1};
 }
 
-SPN_Span SPN_subspan(SPN_Span src, uint32_t begin_idx, uint32_t len) {
+SPN_Span SPN_subspan(SPN_Span src, size_t begin_idx, size_t len) {
   if(begin_idx > src.len) begin_idx = src.len;
   if(begin_idx + len > src.len) len = (src.len - begin_idx);
   return (SPN_Span){.begin = SPN_get(src, begin_idx), .len = len, .element_size = src.element_size};
@@ -36,22 +36,22 @@ bool SPN_contains_subspan(SPN_Span span, SPN_Span subspan) {
   // NOTE this could probably be faster if we did our own comparisons
   const size_t subspan_size_bytes = SPN_get_size_in_bytes(subspan);
 
-  for(uint32_t i = 0; i <= (span.len - subspan.len); i++) {
+  for(size_t i = 0; i <= (span.len - subspan.len); i++) {
     if(memcmp(SPN_get(span, i), subspan.begin, subspan_size_bytes) == 0) return true;
   }
 
   return false;
 }
 
-STAT_Val SPN_find(SPN_Span span, const void * element, uint32_t * o_idx) {
+STAT_Val SPN_find(SPN_Span span, const void * element, size_t * o_idx) {
   return SPN_find_at(span, element, 0, o_idx);
 }
 
-STAT_Val SPN_find_at(SPN_Span span, const void * element, uint32_t at_idx, uint32_t * o_idx) {
+STAT_Val SPN_find_at(SPN_Span span, const void * element, size_t at_idx, size_t * o_idx) {
   if(!is_valid(span)) return LOG_STAT(STAT_ERR_ARGS, "span not valid");
   if(element == NULL) return LOG_STAT(STAT_ERR_ARGS, "element is NULL");
 
-  for(uint32_t i = at_idx; i < span.len; i++) {
+  for(size_t i = at_idx; i < span.len; i++) {
     if(memcmp(SPN_get(span, i), element, span.element_size) == 0) {
       if(o_idx != NULL) *o_idx = i;
       return OK;
@@ -61,22 +61,19 @@ STAT_Val SPN_find_at(SPN_Span span, const void * element, uint32_t at_idx, uint3
   return STAT_OK_NOT_FOUND;
 }
 
-STAT_Val SPN_find_reverse(SPN_Span span, const void * element, uint32_t * o_idx) {
+STAT_Val SPN_find_reverse(SPN_Span span, const void * element, size_t * o_idx) {
   if(span.len == 0) return STAT_OK_NOT_FOUND;
   return SPN_find_reverse_at(span, element, span.len - 1, o_idx);
 }
 
-STAT_Val SPN_find_reverse_at(SPN_Span     span,
-                             const void * element,
-                             uint32_t     at_idx,
-                             uint32_t *   o_idx) {
+STAT_Val SPN_find_reverse_at(SPN_Span span, const void * element, size_t at_idx, size_t * o_idx) {
   if(!is_valid(span)) return LOG_STAT(STAT_ERR_ARGS, "span not valid");
   if(element == NULL) return LOG_STAT(STAT_ERR_ARGS, "element is NULL");
   if(span.len == 0) return STAT_OK_NOT_FOUND;
   if(at_idx >= span.len) at_idx = (span.len - 1);
 
   // we start at +1 and then decrement at the start of the iteration
-  uint32_t i = at_idx + 1;
+  size_t i = at_idx + 1;
   do {
     i--;
     if(memcmp(SPN_get(span, i), element, span.element_size) == 0) {
@@ -88,11 +85,11 @@ STAT_Val SPN_find_reverse_at(SPN_Span     span,
   return STAT_OK_NOT_FOUND;
 }
 
-STAT_Val SPN_find_subspan(SPN_Span span, SPN_Span subspan, uint32_t * o_idx) {
+STAT_Val SPN_find_subspan(SPN_Span span, SPN_Span subspan, size_t * o_idx) {
   return SPN_find_subspan_at(span, subspan, 0, o_idx);
 }
 
-STAT_Val SPN_find_subspan_at(SPN_Span span, SPN_Span subspan, uint32_t at_idx, uint32_t * o_idx) {
+STAT_Val SPN_find_subspan_at(SPN_Span span, SPN_Span subspan, size_t at_idx, size_t * o_idx) {
   if(!is_valid(span)) return LOG_STAT(STAT_ERR_ARGS, "span not valid");
   if(!is_valid(subspan)) return LOG_STAT(STAT_ERR_ARGS, "subspan not valid");
   if(span.element_size != subspan.element_size) {
@@ -102,7 +99,7 @@ STAT_Val SPN_find_subspan_at(SPN_Span span, SPN_Span subspan, uint32_t at_idx, u
 
   const size_t subsp_size_bytes = SPN_get_size_in_bytes(subspan);
 
-  for(uint32_t i = at_idx; i <= (span.len - subspan.len); i++) {
+  for(size_t i = at_idx; i <= (span.len - subspan.len); i++) {
     if(memcmp(SPN_get(span, i), subspan.begin, subsp_size_bytes) == 0) {
       if(o_idx != NULL) *o_idx = i;
       return OK;
@@ -112,7 +109,7 @@ STAT_Val SPN_find_subspan_at(SPN_Span span, SPN_Span subspan, uint32_t at_idx, u
   return STAT_OK_NOT_FOUND;
 }
 
-STAT_Val SPN_find_subspan_reverse(SPN_Span span, SPN_Span subspan, uint32_t * o_idx) {
+STAT_Val SPN_find_subspan_reverse(SPN_Span span, SPN_Span subspan, size_t * o_idx) {
   if(!is_valid(span)) return LOG_STAT(STAT_ERR_ARGS, "span not valid");
   if(!is_valid(subspan)) return LOG_STAT(STAT_ERR_ARGS, "subspan not valid");
   if(span.element_size != subspan.element_size) {
@@ -128,10 +125,10 @@ STAT_Val SPN_find_subspan_reverse(SPN_Span span, SPN_Span subspan, uint32_t * o_
   return SPN_find_subspan_reverse_at(span, subspan, (span.len - 1), o_idx);
 }
 
-STAT_Val SPN_find_subspan_reverse_at(SPN_Span   span,
-                                     SPN_Span   subspan,
-                                     uint32_t   at_idx,
-                                     uint32_t * o_idx) {
+STAT_Val SPN_find_subspan_reverse_at(SPN_Span span,
+                                     SPN_Span subspan,
+                                     size_t   at_idx,
+                                     size_t * o_idx) {
   if(!is_valid(span)) return LOG_STAT(STAT_ERR_ARGS, "span not valid");
   if(!is_valid(subspan)) return LOG_STAT(STAT_ERR_ARGS, "subspan not valid");
   if(span.element_size != subspan.element_size) {
@@ -144,7 +141,7 @@ STAT_Val SPN_find_subspan_reverse_at(SPN_Span   span,
   const size_t subsp_size_bytes = SPN_get_size_in_bytes(subspan);
 
   // we start at +1 and then decrement at the start of the iteration
-  uint32_t i = at_idx + 1;
+  size_t i = at_idx + 1;
   do {
     i--;
     if(memcmp(SPN_get(span, i), subspan.begin, subsp_size_bytes) == 0) {
